+++ HikvisionTimeLapse.Core/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace HikvisionTimeLapse.Core;
-
-public class Class1
-{
-
-}